src/webload/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
kafka-python==2.0.2
jinja2==3.1.3
prometheus-client==0.19.0